Extract host functions list to separate module
