+++ lib/src/executor/host.rs
@@ -202,6 +202,7 @@ use alloc::{
     vec::Vec,
 };
 use core::{fmt, hash::Hasher as _, iter, str};
+use functions::HostFunction;
 
 pub mod runtime_version;
 
@@ -213,6 +214,7 @@ pub use trie::TrieEntryVersion;
 pub use vm::HeapPages;
 pub use zstd::Error as ModuleFormatError;
 
+mod functions;
 mod tests;
 mod zstd;
 
@@ -3826,435 +3828,6 @@ pub enum Error {
         /// Name of the function being called.
         function: &'static str,
     },
-}
-
-macro_rules! externalities {
-    ($($ext:ident,)*) => {
-        /// List of possible externalities.
-        #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-        #[allow(non_camel_case_types)]
-        enum HostFunction {
-            $(
-                $ext,
-            )*
-        }
-
-        impl HostFunction {
-            fn by_name(name: &str) -> Option<Self> {
-                $(
-                    if name == stringify!($ext) {
-                        return Some(HostFunction::$ext);
-                    }
-                )*
-                None
-            }
-
-            fn name(&self) -> &'static str {
-                match self {
-                    $(
-                        HostFunction::$ext => stringify!($ext),
-                    )*
-                }
-            }
-        }
-    };
-}
-
-externalities! {
-    ext_storage_set_version_1,
-    ext_storage_get_version_1,
-    ext_storage_read_version_1,
-    ext_storage_clear_version_1,
-    ext_storage_exists_version_1,
-    ext_storage_clear_prefix_version_1,
-    ext_storage_clear_prefix_version_2,
-    ext_storage_root_version_1,
-    ext_storage_root_version_2,
-    ext_storage_changes_root_version_1,
-    ext_storage_next_key_version_1,
-    ext_storage_append_version_1,
-    ext_storage_start_transaction_version_1,
-    ext_storage_rollback_transaction_version_1,
-    ext_storage_commit_transaction_version_1,
-    ext_default_child_storage_get_version_1,
-    ext_default_child_storage_read_version_1,
-    ext_default_child_storage_storage_kill_version_1,
-    ext_default_child_storage_storage_kill_version_2,
-    ext_default_child_storage_storage_kill_version_3,
-    ext_default_child_storage_clear_prefix_version_1,
-    ext_default_child_storage_clear_prefix_version_2,
-    ext_default_child_storage_set_version_1,
-    ext_default_child_storage_clear_version_1,
-    ext_default_child_storage_exists_version_1,
-    ext_default_child_storage_next_key_version_1,
-    ext_default_child_storage_root_version_1,
-    ext_default_child_storage_root_version_2,
-    ext_crypto_ed25519_public_keys_version_1,
-    ext_crypto_ed25519_generate_version_1,
-    ext_crypto_ed25519_sign_version_1,
-    ext_crypto_ed25519_verify_version_1,
-    ext_crypto_ed25519_batch_verify_version_1,
-    ext_crypto_sr25519_public_keys_version_1,
-    ext_crypto_sr25519_generate_version_1,
-    ext_crypto_sr25519_sign_version_1,
-    ext_crypto_sr25519_verify_version_1,
-    ext_crypto_sr25519_verify_version_2,
-    ext_crypto_sr25519_batch_verify_version_1,
-    ext_crypto_ecdsa_generate_version_1,
-    ext_crypto_ecdsa_sign_version_1,
-    ext_crypto_ecdsa_public_keys_version_1,
-    ext_crypto_ecdsa_verify_version_1,
-    ext_crypto_ecdsa_verify_version_2,
-    ext_crypto_ecdsa_sign_prehashed_version_1,
-    ext_crypto_ecdsa_verify_prehashed_version_1,
-    ext_crypto_ecdsa_batch_verify_version_1,
-    ext_crypto_secp256k1_ecdsa_recover_version_1,
-    ext_crypto_secp256k1_ecdsa_recover_version_2,
-    ext_crypto_secp256k1_ecdsa_recover_compressed_version_1,
-    ext_crypto_secp256k1_ecdsa_recover_compressed_version_2,
-    ext_crypto_start_batch_verify_version_1,
-    ext_crypto_finish_batch_verify_version_1,
-    ext_hashing_keccak_256_version_1,
-    ext_hashing_keccak_512_version_1,
-    ext_hashing_sha2_256_version_1,
-    ext_hashing_blake2_128_version_1,
-    ext_hashing_blake2_256_version_1,
-    ext_hashing_twox_64_version_1,
-    ext_hashing_twox_128_version_1,
-    ext_hashing_twox_256_version_1,
-    ext_offchain_index_set_version_1,
-    ext_offchain_index_clear_version_1,
-    ext_offchain_is_validator_version_1,
-    ext_offchain_submit_transaction_version_1,
-    ext_offchain_network_state_version_1,
-    ext_offchain_timestamp_version_1,
-    ext_offchain_sleep_until_version_1,
-    ext_offchain_random_seed_version_1,
-    ext_offchain_local_storage_set_version_1,
-    ext_offchain_local_storage_compare_and_set_version_1,
-    ext_offchain_local_storage_get_version_1,
-    ext_offchain_local_storage_clear_version_1,
-    ext_offchain_http_request_start_version_1,
-    ext_offchain_http_request_add_header_version_1,
-    ext_offchain_http_request_write_body_version_1,
-    ext_offchain_http_response_wait_version_1,
-    ext_offchain_http_response_headers_version_1,
-    ext_offchain_http_response_read_body_version_1,
-    ext_trie_blake2_256_root_version_1,
-    ext_trie_blake2_256_root_version_2,
-    ext_trie_blake2_256_ordered_root_version_1,
-    ext_trie_blake2_256_ordered_root_version_2,
-    ext_trie_keccak_256_root_version_1,
-    ext_trie_keccak_256_root_version_2,
-    ext_trie_keccak_256_ordered_root_version_1,
-    ext_trie_keccak_256_ordered_root_version_2,
-    ext_trie_blake2_256_verify_proof_version_1,
-    ext_trie_blake2_256_verify_proof_version_2,
-    ext_trie_keccak_256_verify_proof_version_1,
-    ext_trie_keccak_256_verify_proof_version_2,
-    ext_misc_print_num_version_1,
-    ext_misc_print_utf8_version_1,
-    ext_misc_print_hex_version_1,
-    ext_misc_runtime_version_version_1,
-    ext_allocator_malloc_version_1,
-    ext_allocator_free_version_1,
-    ext_logging_log_version_1,
-    ext_logging_max_level_version_1,
-}
-
-impl HostFunction {
-    // TODO: make this a `const fn` function
-    fn signature(&self) -> vm::Signature {
-        match *self {
-            HostFunction::ext_storage_set_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_storage_get_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_read_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_clear_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_storage_exists_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_storage_clear_prefix_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_storage_clear_prefix_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_root_version_1 => crate::signature!(() => vm::ValueType::I64),
-            HostFunction::ext_storage_root_version_2 => {
-                crate::signature!((vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_changes_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_next_key_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_storage_append_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_storage_start_transaction_version_1 => crate::signature!(() => ()),
-            HostFunction::ext_storage_rollback_transaction_version_1 => crate::signature!(() => ()),
-            HostFunction::ext_storage_commit_transaction_version_1 => crate::signature!(() => ()),
-            HostFunction::ext_default_child_storage_get_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_read_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_storage_kill_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_default_child_storage_storage_kill_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_default_child_storage_storage_kill_version_3 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_clear_prefix_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_default_child_storage_clear_prefix_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_set_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_default_child_storage_clear_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_default_child_storage_exists_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_default_child_storage_next_key_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_default_child_storage_root_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ed25519_public_keys_version_1 => {
-                crate::signature!((vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ed25519_generate_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ed25519_sign_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ed25519_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ed25519_batch_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_sr25519_public_keys_version_1 => {
-                crate::signature!((vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_sr25519_generate_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_sr25519_sign_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_sr25519_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_sr25519_verify_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_sr25519_batch_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ecdsa_generate_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ecdsa_sign_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ecdsa_public_keys_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ecdsa_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ecdsa_verify_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ecdsa_sign_prehashed_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_ecdsa_verify_prehashed_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_ecdsa_batch_verify_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_crypto_secp256k1_ecdsa_recover_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_secp256k1_ecdsa_recover_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_secp256k1_ecdsa_recover_compressed_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_secp256k1_ecdsa_recover_compressed_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I32) => vm::ValueType::I64)
-            }
-            HostFunction::ext_crypto_start_batch_verify_version_1 => crate::signature!(() => ()),
-            HostFunction::ext_crypto_finish_batch_verify_version_1 => {
-                crate::signature!(() => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_keccak_256_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_keccak_512_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_sha2_256_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_blake2_128_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_blake2_256_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_twox_64_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_twox_128_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_hashing_twox_256_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_offchain_index_set_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_offchain_index_clear_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_offchain_is_validator_version_1 => {
-                crate::signature!(() => vm::ValueType::I32)
-            }
-            HostFunction::ext_offchain_submit_transaction_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_network_state_version_1 => {
-                crate::signature!(() => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_timestamp_version_1 => {
-                crate::signature!(() => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_sleep_until_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_offchain_random_seed_version_1 => {
-                crate::signature!(() => vm::ValueType::I32)
-            }
-            HostFunction::ext_offchain_local_storage_set_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_offchain_local_storage_compare_and_set_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_offchain_local_storage_get_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_local_storage_clear_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_offchain_http_request_start_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_http_request_add_header_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_http_request_write_body_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_http_response_wait_version_1 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_http_response_headers_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_offchain_http_response_read_body_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_trie_blake2_256_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_blake2_256_root_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_blake2_256_ordered_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_blake2_256_ordered_root_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_root_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_ordered_root_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_ordered_root_version_2 => {
-                crate::signature!((vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_blake2_256_verify_proof_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_blake2_256_verify_proof_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_verify_proof_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64) => vm::ValueType::I32)
-            }
-            HostFunction::ext_trie_keccak_256_verify_proof_version_2 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I64, vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_misc_print_num_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_misc_print_utf8_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_misc_print_hex_version_1 => {
-                crate::signature!((vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_misc_runtime_version_version_1 => {
-                crate::signature!((vm::ValueType::I64) => vm::ValueType::I64)
-            }
-            HostFunction::ext_allocator_malloc_version_1 => {
-                crate::signature!((vm::ValueType::I32) => vm::ValueType::I32)
-            }
-            HostFunction::ext_allocator_free_version_1 => {
-                crate::signature!((vm::ValueType::I32) => ())
-            }
-            HostFunction::ext_logging_log_version_1 => {
-                crate::signature!((vm::ValueType::I32, vm::ValueType::I64, vm::ValueType::I64) => ())
-            }
-            HostFunction::ext_logging_max_level_version_1 => {
-                crate::signature!(() => vm::ValueType::I32)
-            }
-        }
-    }
 }
 
 // Glue between the `allocator` module and the `vm` module.
